default img will dlt together when dlt customer

+++ project/customers_delete.php
@@ -1,11 +1,11 @@
 <?php
 // include database connection
 include 'config/database.php';
-try {     
+try {
     // get record ID
     // isset() is a PHP function used to verify if a value is there or not
-    $id=isset($_GET['id']) ? $_GET['id'] :  die('ERROR: Record ID not found.');
-    $select ="SELECT user AS customername, image FROM customers where id=:id";
+    $id = isset($_GET['id']) ? $_GET['id'] :  die('ERROR: Record ID not found.');
+    $select = "SELECT user AS customername, image FROM customers where id=:id";
     $stmt = $con->prepare($select);
     $stmt->bindParam(":id", $id);
     $stmt->execute();
@@ -13,31 +13,34 @@ try {
     extract($row);
 
     // delete query
-    $check ="SELECT username FROM order_summary WHERE username=:username";
+    $check = "SELECT username FROM order_summary WHERE username=:username";
     $stmt = $con->prepare($check);
     $stmt->bindParam(":username", $customername);
     $stmt->execute();
     $count = $stmt->rowCount();
-    
-    if($count > 0){
+
+    if ($count > 0) {
         echo "This customer has order so cannot be delete.";
-    }else{
+    } else {
         $query = "DELETE FROM customers WHERE id = ?";
         $stmt = $con->prepare($query);
         $stmt->bindParam(1, $id);
-         
-        if($stmt->execute()){
-            unlink("uploads/" . $row['image']);
+
+        if ($stmt->execute()) {
+            if ($row['image'] != "nonprofile2.png") {
+                unlink("uploads/" . $row['image']);
+                
+            }header('Location: customers_read.php?action=deleted');
+    
             // redirect to read records page and
             // tell the user record was deleted
-            header('Location: customers_read.php?action=deleted');
-        }else{
+           
+        } else {
             die('Unable to delete record.');
         }
     }
-    
 }
 // show error
-catch(PDOException $exception){
+catch (PDOException $exception) {
     die('ERROR: ' . $exception->getMessage());
 }
